Align Flash GQA tests with dtype contracts
FlashAttention is intentionally refused for FP32 after removing hidden precision conversion. Exercise the non-expanding GQA branch with FP16 inputs so the tests validate the supported path.\n\nCo-Authored-By: Codex <codex@openai.com>

+++ backend/tests/krea2_gqa_dispatch_test.py
@@ -64,9 +64,9 @@ class Krea2GQADispatchTest(unittest.TestCase):
         self.assertFalse(capture["enable_gqa"])
 
     def test_flash_backend_leaves_kv_unexpanded(self):
-        attn = _attn()
+        attn = _attn().half()
         attn._attn_backend = "flash"
-        hidden = torch.randn(1, 5, attn.hidden_size)
+        hidden = torch.randn(1, 5, attn.hidden_size, dtype=torch.float16)
 
         capture = {}
         original = BACKENDS["flash"]
